# api_gateway/api/docs/swagger.yaml
definitions:
...
  user_service.Couriers:
    properties:
      active:
        type: boolean
      branch_id:
        type: integer
      created_at:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      max_order_count:
        type: integer
      password:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  user_service.CreateCouriersRequest:
    properties:
      branch_id:
        type: integer
      firstname:
        type: string
      lastname:
        type: string
      login:
        type: string
      max_order_count:
        type: integer
      password:
        type: string
      phone:
        type: string
    type: object
...
  user_service.ListCouriersResponse:
    properties:
      count:
        type: integer
      couriers:
        items:
          $ref: '#/definitions/user_service.Couriers'
        type: array
    type: object
...
  user_service.Response:
    properties:
      message:
        type: string
    type: object
...
  user_service.UpdateCouriersRequest:
    properties:
      branch_id:
        type: integer
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      login:
        type: string
      max_order_count:
        type: integer
      password:
        type: string
      phone:
        type: string
    type: object
...
  v1.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /v1/courier/create:
    post:
      consumes:
      - application/json
      description: Create a new courier with the provided details
      parameters:
      - description: data of the courier
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateCouriersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user_service.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create a new courier
      tags:
      - courier
  /v1/courier/delete/{id}:
    delete:
      consumes:
      - application/json
      description: delete a courier by its unique identifier
      parameters:
      - description: Catgeory ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete a Catgory
      tags:
      - courier
  /v1/courier/get/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a courier by its unique identifier
      parameters:
      - description: Courier ID to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.Couriers'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get a courier by ID
      tags:
      - courier
  /v1/courier/list:
    get:
      consumes:
      - application/json
      description: get courier
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        type: integer
      - default: 1
        description: page for response
        in: query
        name: page
        type: integer
      - description: firstname
        in: query
        name: firstname
        type: string
      - description: lastname
        in: query
        name: lastname
        type: string
      - description: phone
        in: query
        name: phone
        type: string
      - description: search by created_at_from
        in: query
        name: created_at_from
        type: string
      - description: search by created_at_to
        in: query
        name: created_at_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user_service.ListCouriersResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: GetAll Courier
      tags:
      - courier
  /v1/courier/update/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing courier with the provided details
      parameters:
      - description: Courier ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Updated data for the courier
        in: body
        name: courier
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateCouriersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_service.UpdateCouriersRequest'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/v1.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update an existing courier
      tags:
      - courier
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"